pyproject.toml
[project]
name = "rag-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "faiss-cpu>=1.13.0",
    "google-genai>=1.51.0",
    "pymilvus>=2.6.4",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.1",
    "transformers>=4.57.1",
]
